Factored start search code out into a new controller class
