+++ client/src/main/resources/views/workspace/StartSearchView.fxml
@@ -19,7 +19,7 @@
 <Tab text="   Start Search" closable="false"
      xmlns="http://javafx.com/javafx"
      xmlns:fx="http://javafx.com/fxml"
-     fx:controller="org.iconic.workspace.WorkspaceController">
+     fx:controller="org.iconic.project.search.StartSearchController">
     <!-- Define the primary screen so we can use percentage based widths -->
     <fx:define>
         <Screen fx:factory="getPrimary" fx:id="screen"/>
